+++ CHANGELOG.md
@@ -1,4 +1,3 @@
 > Read detailed change logs in https://github.com/sakurawald/fuji/commits/dev/
 
-- (MC 1.20 ~ MC 1.20.1) [core] fix: should not display the `closing tag` for `custom color tags` (e.g. `<#FF0000>`)
-  - This greatly improve the readability of `in-game inspector` and user experience.
+- [economy & leaderboard] feature: improve the compatibility with async read/write operations.